feat(fancy-bar): add battery indicator icon
This took quite a bit of tinkering to get just right, but I'm happy with
the results.

I used the ii Quickshell config[1] as inspiration, but got rid of and
changed most of its battery indicator code. I just want a simple icon
with an internal progress bar, and a percentage overlay.

This also looks better since it uses a battery icon, whereas the ii
config uses a rounded progress bar with a Material icon next to it.

For the icon, I initially started with an SVG, but that was problematic
since SVGs usually have a larger viewport than the icon itself, so it
was painful getting the progress bar to fit precisely inside it. I
considered changing the SVG itself, but then wondered if it could be
drawn with QtQuick itself. Turns out that it can, and Claude Sonnet 4.5
did the implementation. This looks and works *much* better than the SVG,
since I can style it just right (changing the stroke width, etc.), and
can also precisely determine its boundaries for clipping the progress
bar inside it.

I tried all possible variations and sizes of the icon, progress bar and
text, and it all looked pixel-perfect to me. :)

Next, I want to add a lightning bolt indicator next to the battery icon while
charging, and some kind of popup to show time left, etc.

[1]: https://github.com/end-4/dots-hyprland/blob/449df7f161e6435569bc7d9499b2e444dd8aa153/dots/.config/quickshell/ii/