feat(fancy-bar): add battery icon shimmer and pulse when charging
So this took a *ridiculous* amount of time to get right...

Initially I just thought I'd add a lightning icon next to the battery
icon to indicate charging, but I found that that would either clutter up
my status bar, which is quite short (30px), or would make the lightning
icon illegible, which defeats the purpose.

So after some initial design chat with Claude Sonnet 4.5, I settled on a
subtle gradient "shimmer" from left to right (or bottom to top in
vertical orientation), and a subtle hue pulse, to indicate charging.

This is much more visually pleasing, but it was a *bitch* to get just
how I wanted it. It required *a lot* of back and forth with Claude, and
*a lot* of manual tweaking and testing, and after a few tedious hours, I
finally arrived at something I'm happy with. Perfectionism is a curse...

QtQuick is very flexible and makes these effects relatively simple to
implement, but getting pixel-perfect positionining at different sizes
and orientations was an absolute nightmare.

That said, I'm fairly certain this is the nicest looking battery
indicator I've ever seen. Even if it only takes 700 pixels on my screen,
I hope I'll be happily staring at it for years to come. :)