feat(fancy-bar): add subtle drop shadow to bar
This took quite a bit of tinkering, but I made it... The tricky thing was
figuring out the positioning, as usual, but also a way for the shadow to
not add a gap between windows, and remain in the background. The
breakthrough was discovering `WlrLayershell`.

Then I spent a frustrating amount of time getting rid of a barely
visible cutoff line for the shadow, which looked jarring on light
backgrounds. It turns out that it's most visible with `blurMax`, and not
`blurMultiplier`, so I ended up using the latter.

It finally looks nice, and is not too distracting. This shadow is
deliberately stronger and more "grounded" than the shadow on niri
windows. I wanted it to look more like part of the desktop itself,
whereas windows can "float" higher above it.