feat(fancy-bar): add RAM widget
This adds a RAM widget showing a breakdown of the current RAM usage. It
follows the calculations used by `free(1)`.

I considered several approaches for rendering the main graph, such as
stacked bars or a pie chart, but ultimately settled on this single
proportional color-coded bar. It makes it easy to see at a glance the
overall usage, as well as the usage of individual components.

Calculating exact RAM usage on Linux is an inexact science, considering
the different definitions of used/buffered/cached, so the percentages of
total RAM are not exactly precise. But it's relatively useful in any
case, so I'm happy with how this turned out.

Grok Code Fast 1 was useful with generating the initial versions of
this, which was easier this time around since I could use the CPU widget
as reference. As usual, everything was carefully reviewed, and I
manually changed a lot of it.

Like with the CPU widget, this could possibly be done more efficiently
to avoid frequent polling, but the performance overhead is negligible on
my 8-year-old laptop, so this is fine for now.