feat(fancy-bar): add network widget
This adds a widget to monitor network information, such as current
upload/download rates in both graph and numerical form, LAN and WAN IPs,
and information specific to the network type (e.g. SSID and signal
strength for WiFi). It allows switching between interfaces via a context
menu, and shows a dynamic icon, which for wireless interfaces also
indicates the signal strength via the number of visible bars.

It is the most complex widget I've added so far, and I think the
complexity is worth it. :)

Overall I'm quite pleased with how QtQuick and Quickshell make this type
of advanced visualization and UI relatively simple to implement. I had
some issues getting everything aligned properly, but managed to get it
as close to pixel-perfect as possible.

Grok Code Fast 1 and Kimi K2 Thinking were helpful with design
discussion and prototyping, but as usual, everything was reviewed,
tested, and largely tweaked by myself.