feat(fancy-bar): slightly nicer workspaces widget
This rounds off the workspaces widget entirely instead of only having
rounded bottom corners, and centers it vertically in the bar. This is a
bit cleaner visually if the bar is placed on the bottom, and I like the
symmetry more.

It also removes the `visible` hack to limit the number of shown
workspaces, and uses the new `maxCount` property exposed by the
`Niri.workspaces` model, which ensures that the width is properly
calculated. This requires v0.1.1 of qml-niri (ffe3f9e).

It also adds a short animation when switching workspaces, and marks the
workspaces with windows with a slightly darker color than the inactive
one. The different color highlights which workspaces have windows, and
which don't.

I'm far from happy with this widget, and have ideas to make it fancier
in the future, but this will have to do for now.