fix(fancy-bar): correct bar height with niri v25.11
niri v25.11 updated smithay[1], which seems to implement a different
mechanism for shell layer positioning[2]. This caused my `barShadow`
component to be rendered *below* the `bar`, making the bar appear double
in height (60px vs 30px).

I'm not sure if this is a bug or not, and where the bug actually is, or
if it's an improved Wayland protocol implementation, and this is how
things will work from now on... but I confirmed that this change renders
the bar correctly with both niri v25.08 and v25.11, and Quickshell
v0.2.1, so I'll keep it.

Honestly, the layers of complexity are insane, and I have better things
to do than understand this madness.

[1]: https://github.com/Smithay/smithay

[2]: From niri's changelog[3]:

> Added support for layer-shell v5 with its set_exclusive_edge() request,
and changed layer surface ordering to consider exclusive zone surfaces
first.

[3]: https://github.com/YaLTeR/niri/discussions/2917